.github/workflows/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
tensorflow==2.15.0
python-multipart==0.0.9
jinja2==3.1.4
numpy==1.26.4
pillow==10.4.0
sqlalchemy==2.0.35